extract: logic from DetailStep
The component has been divided into constituent subcomponents -
heading and input fields. The logic is still kept in the InputFields
component as creating context store is overkill for this case.
The logic will be extracted into use-case function in a later commit.

Validation schema logic has been extracted out of templateHelpers
and placed into validateUserInput file inside the folder.
