+++ src/features/createAndEditTemplate/describeTemplate/core/domain/validateUserInput.ts
@@ -1,0 +1,6 @@
+import * as Yup from 'yup';
+
+export const TemplateValidationSchema = Yup.object().shape({
+  name: Yup.string().max(255, 'Too Long!').required('Required'),
+  description: Yup.string().max(255, 'Too Long!'),
+});
